Speed up jet recombination
When removing a pseudojet from the grid (because two have been recombined, or
because it has been promoted to jet) simply reassign its index in the grid,
instead of removing the original jet and adding the jet in the new position.

Saves 0.8 ms/ev (tested on Dell XPS notebook with a GeForce GTX 960M).