Increase the cell size to have an integer number of cells in phi
For example, for a jet cone size of 0.4, using a grid cell size of 0.4x0.4
results in 15.708... cells - that is 16 cells, the last one of which is only
partially full, which in turn requires special handling of the neighbouring
cells around the 0 <-> 2pi transition.

To have an integer number of cells in phi, we increase the cell size to
2 pi / 15, or 0.419...

The increase in cell size seems negligible from the occupancy point of view.
The removal of the special cases improves the reconstruction time by ~1% .